move header content into a separate file
