+++ components/common/page-layout/Sidebar.tsx
@@ -32,7 +32,7 @@ import Link from '../Link';
 import ModalContainer from '../ModalContainer';
 import WorkspaceAvatar from '../WorkspaceAvatar';
 import CreateWorkspaceForm from './CreateWorkspaceForm';
-import Header from './Header';
+import { headerHeight } from './Header';
 import PageNavigation from './PageNavigation';
 import NewPageMenu from '../NewPageMenu';
 
@@ -73,6 +73,25 @@ const SidebarContainer = styled.div`
     opacity: 1;
   }
 `;
+
+const SidebarHeader = styled.div(({ theme }) => ({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  padding: theme.spacing(0, 2),
+  '& .MuiIconButton-root': {
+    opacity: 0,
+    transition: theme.transitions.create('opacity', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  '&:hover .MuiIconButton-root': {
+    opacity: 1
+  },
+  // necessary for content to be below app bar
+  minHeight: headerHeight
+}));
 
 interface SidebarProps {
   closeSidebar: () => void;
@@ -197,25 +216,13 @@ export default function Sidebar ({ closeSidebar, favorites }: SidebarProps) {
       <Box display='flex' flexDirection='column' sx={{ height: '100%', flexGrow: 1 }}>
         <Box sx={{ flexGrow: 1 }}>
           <Box display='flex' flexDirection='column' sx={{ height: '100%' }}>
-            <Header>
+            <SidebarHeader>
               <Typography><strong>{space.name}</strong></Typography>
               <IconButton onClick={closeSidebar}>
                 <ChevronLeftIcon />
               </IconButton>
-            </Header>
+            </SidebarHeader>
             <Divider sx={{ mb: 3 }} />
-            {/* <Box>
-            <List>
-              <NextLink href='' passHref>
-                <ListItem button component='a' disableRipple
-                  sx={{ py: 1, color: greyColor + ' !important' }}>
-                  <ListItemText disableTypography>
-                      <Box sx={{ fontSize: 14, fontWeight: 500 }}>Settings</Box>
-                  </ListItemText>
-                </ListItem>
-              </NextLink>
-            </List>
-          </Box> */}
             {favoritePageIds.length > 0 && (
               <Box mb={2}>
                 <Typography sx={{ color: greyColor2, fontSize: 12, letterSpacing: '0.03em', fontWeight: 600, px: 2, mb: 0.5 }}>
